// src/app/core/models/catalogs/domicile.ts
export class IDomicile {
  id?: number;
  cveState?: number | null;
  cveMunicipality?: string;
  cveLocality?: string;
  numExt?: string;
  numInt?: string;
  code?: string;
  cveVia1?: string;
  cveVia2?: string;
  cveVia3?: string;
  description?: string;
  latitude?: string;
  longitude?: string;
  viaName?: string;
  viaOrigin?: string;
  viaDestination?: string;
  viaChaining?: string;
  userCreation?: string;
  creationDate?: string;
  modificationUser?: string;
  modificationDate?: string;
  cveSettlement?: string;
  version?: number | null;
  idRequest?: number | null;
  aliasWarehouse?: string;
  idDelegation?: number | null;
}